% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp.R
\name{bb}
\alias{bb}
\title{Bollinger Bands}
\usage{
bb(price, n, sd = 2)
}
\arguments{
\item{price}{A numeric vector of prices}

\item{n}{The period for the moving average}

\item{sd}{The number of standard deviations for the bands}
}
\value{
A list containing the lower band, moving average, upper band, and percentage B
}
\description{
Bollinger Bands
}
