% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{collapse}
\alias{collapse}
\title{Collapse a vector into a single string}
\usage{
collapse(x)
}
\arguments{
\item{x}{A vector of elements to collapse.}
}
\value{
A single string with all elements from the vector separated by a comma and a space.
}
\description{
This function collapses a vector into a single string, with elements separated by a comma and a space.
}
\examples{
collapse(c("a", "b", "c"))
# Returns: "a, b, c"
}
