% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp.R
\name{ema}
\alias{ema}
\title{Exponential Moving Average}
\usage{
ema(price, n, wilder = FALSE)
}
\arguments{
\item{price}{A numeric vector of prices}

\item{n}{The period for the EMA}

\item{wilder}{Whether to use Wilder's smoothing}
}
\value{
A numeric vector containing the EMA values
}
\description{
Exponential Moving Average
}
