% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feature_counts}
\alias{feature_counts}
\title{Feature Counts for Genes}
\format{
A data frame with rows and variables:
\describe{
  \item{GeneID}{Ensembl gene identifier}
  \item{GeneSymbol}{Gene symbol}
  \item{GeneBiotype}{Type of gene}
  \item{B24568, B24560, ...}{Counts for each sample}
}
}
\source{
Gene expression study
}
\usage{
feature_counts
}
\description{
A dataset containing feature counts for various genes across different samples.
}
\keyword{datasets}
