% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fib}
\alias{fib}
\title{Fibonacci Retracement Levels
Calculate the Fibonacci retracement levels for a given price range}
\usage{
fib(high, low)
}
\arguments{
\item{high}{The high price}

\item{low}{The low price}
}
\value{
A list containing the retracement levels and prices
}
\description{
Fibonacci Retracement Levels
Calculate the Fibonacci retracement levels for a given price range
}
