% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ichimoku_cloud}
\alias{ichimoku_cloud}
\title{Ichimoku Cloud
Calculate the Ichimoku Cloud for a given price series}
\usage{
ichimoku_cloud(
  high,
  low,
  close,
  tenkan_period = 9L,
  kijun_period = 26L,
  senkou_period = 52L
)
}
\arguments{
\item{high}{The high prices}

\item{low}{The low prices}

\item{close}{The closing prices}

\item{tenkan_period}{The period for the Tenkan-sen (Conversion Line)}

\item{kijun_period}{The period for the Kijun-sen (Base Line)}

\item{senkou_period}{The period for the Senkou Span B (Leading Span B)}
}
\value{
A list containing the Ichimoku Cloud components
}
\description{
Ichimoku Cloud
Calculate the Ichimoku Cloud for a given price series
}
