% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.empty.R
\name{is.empty.default}
\alias{is.empty.default}
\title{Default method for is.empty}
\usage{
\method{is.empty}{default}(x)
}
\arguments{
\item{x}{An R object}
}
\value{
This function always throws an error
}
\description{
This function is called when is.empty is applied to an object of a class
for which no specific method is defined.
}
\examples{
\dontrun{
# This will throw an error
is.empty(quote(x))
}
}
