% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.empty.R
\name{is.empty.factor}
\alias{is.empty.factor}
\title{Check if a factor is empty}
\usage{
\method{is.empty}{factor}(x)
}
\arguments{
\item{x}{A factor object}
}
\value{
A logical value: TRUE if the factor is empty, FALSE otherwise
}
\description{
This function checks if a factor has length zero.
}
\examples{
fct1 <- factor(levels = c("A", "B"))
is.empty(fct1)  # Returns TRUE

fct2 <- factor(c("A", "B", "A"))
is.empty(fct2)  # Returns FALSE
}
