% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp.R
\name{macd}
\alias{macd}
\title{Moving Average Convergence Divergence (MACD)}
\usage{
macd(price, s, l, k, percent = TRUE)
}
\arguments{
\item{price}{A numeric vector of prices}

\item{s}{The short-term period}

\item{l}{The long-term period}

\item{k}{The signal line period}

\item{percent}{Whether to return the MACD as a percentage}
}
\value{
A list containing the MACD and signal line values
}
\description{
Moving Average Convergence Divergence (MACD)
}
