% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mom}
\alias{mom}
\title{Momentum}
\usage{
mom(price, n)
}
\arguments{
\item{price}{A numeric vector of prices}

\item{n}{The period for momentum calculation}
}
\value{
A numeric vector containing the momentum values
}
\description{
Momentum
}
