% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{printCapture}
\alias{printCapture}
\title{Print and capture the output of an object}
\usage{
printCapture(obj)
}
\arguments{
\item{obj}{The object to print and capture the output from.}
}
\value{
A string representing the printed output of the object.
}
\description{
This function prints an object and captures the output as a string.
}
\details{
This is useful for printing data.frames to log files etc.
}
\examples{
df <- data.frame(a = 1:3, b = letters[1:3])
printCapture(df)
# Returns a string representation of the data frame
}
