% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp.R
\name{roc}
\alias{roc}
\title{Rate of Change (ROC)}
\usage{
roc(price, n, type = "c")
}
\arguments{
\item{price}{A numeric vector of prices}

\item{n}{The period for ROC calculation}

\item{type}{The type of ROC calculation: 'c' for continuous (default) or 'd' for discrete}
}
\value{
A numeric vector containing the ROC values
}
\description{
Rate of Change (ROC)
}
