% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp.R
\name{rsi}
\alias{rsi}
\title{Relative Strength Index (RSI)}
\usage{
rsi(price, n, method = "e")
}
\arguments{
\item{price}{A numeric vector of prices}

\item{n}{The period for RSI calculation}

\item{method}{The method for average calculation: 'e' for EMA (default) or 's' for SMA}
}
\value{
A numeric vector containing the RSI values
}
\description{
Relative Strength Index (RSI)
}
