% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp.R
\name{sma}
\alias{sma}
\title{Simple Moving Average (SMA)}
\usage{
sma(price, n)
}
\arguments{
\item{price}{A numeric vector of prices}

\item{n}{The period for SMA calculation}
}
\value{
A numeric vector containing the SMA values
}
\description{
Simple Moving Average (SMA)
}
