% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_macd.R
\name{stat_macd}
\alias{stat_macd}
\title{Moving average convergence divergence (macd) `ggplot2` layer}
\usage{
stat_macd(
  mapping = NULL,
  data = NULL,
  geom = "line",
  position = "identity",
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE,
  size = 1,
  alpha_lines = 0.75,
  alpha_histogram = 0.5,
  colour_lines = list(macd = "blue", macd_signal = "red"),
  colour_histogram = list(positive = NA_character_, negative = NA_character_),
  fill_histogram = list(positive = "green", negative = "red"),
  ...
)
}
\arguments{
\item{mapping}{A `ggplot2::aes` object (required - default `NULL`).}

\item{data}{A `data.table` object (required - default `NULL`).}

\item{size}{A `numeric` vector of length one; the size of the line (optional - default `1.75`).}

\item{alpha_lines}{A `numeric` vector of length one; the alpha of the lines (optional - default `0.75`).}

\item{alpha_histogram}{A `numeric` vector of length one; the alpha of the histogram (optional - default `0.5`).}

\item{colour_lines}{A named or unnamed `list` of two elements "macd" and "macd_signal" (optional - default `list(macd = "blue", macd_signal = "red")`).}

\item{colour_histogram}{A named or unnamed `list` of two elements "positive" and "negative" (optional - default `list(positive = NA_character_, negative = NA_character_)`).}

\item{fill_histogram}{A named or unnamed `list` of two elements "positive" and "negative" (optional - default `list(positive = "green", negative = "red")`).}
}
\value{
A `ggplot2::layer` object.
}
\description{
`stat_macd` is a `ggplot2` layer that allows you to plot a `ggplot2` layer of moving average convergence divergence (macd) by providing the column names `ggplot2::aes` of the previously calculated metrics.

You are free to use whatever algorithm you desire; the result will be a line plot overlayed on a histogram.

To use this layer provide `ggplot2::aes` values for `x` (datetime x-axis) and `macd` (y-axis) and `macd_signal` (y-axis) and `macd_diff` (y-axis).

See examples for more details.
}
\details{
This is a `ggplot2` extension; it is used with the `+` operator for adding a layer to a `ggplot2` object.
}
\section{Aesthetics}{

\code{stat_movingaverages} understands the following aesthetics (required aesthetics are in bold):
\itemize{
  \item \strong{x} -- datetime (x-axis)
  \item \strong{macd} -- y-axis
  \item \strong{macd_signal} -- y-axis
  \item \strong{macd_diff} -- y-axis
}
}

\author{
Dereck de Mezquita
}
