% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_shade_ranges.R
\name{stat_shade_ranges}
\alias{stat_shade_ranges}
\title{Shaded Date Ranges `ggplot2` layer}
\usage{
stat_shade_ranges(
  mapping = NULL,
  data = NULL,
  geom = "linerange",
  position = "identity",
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{A `ggplot2::aes` object (required - default `NULL`).}

\item{data}{A `data.table` object (required - default `NULL`).}

\item{...}{Additional arguments passed to `ggplot2::layer`.}
}
\value{
A `ggplot2::layer` object.
}
\description{
Shaded Date Ranges `ggplot2` layer
}
\details{
This is a `ggplot2` extension; it is used with the `+` operator for adding a layer to a `ggplot2` object.

You can set the colour by fill aesthetic with a column.
}
\examples{
\dontrun{
dt |>
   ggplot2::ggplot(ggplot2::aes(x = datetime)) +
   ggplot2::geom_line(ggplot2::aes(y = close), linewidth = 1.5, colour = "white") +
   plots$stat_shade_ranges(ggplot2::aes(group = call, fill = call), alpha = 0.25) +
   ggplot2::scale_color_manual(
       name = "call",
       values = c("buy" = "green", "sell" = "red", "none" = "gray"),
       aesthetics = "fill"
  )
}

}
\author{
Dereck Mezquita
}
