% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dereck.R
\name{theme_dereck}
\alias{theme_dereck}
\title{Add derecksnotes theme to ggplot chart}
\usage{
theme_dereck()
}
\description{
Add derecksnotes theme to ggplot chart
}
\examples{
iris |>
    ggplot2::ggplot(ggplot2::aes(x = Sepal.Length, y = Sepal.Width)) +
    ggplot2::geom_point(ggplot2::aes(color = Species, shape = Species)) +
    ggplot2::labs(title = "Iris data", subtitle = "Classic data", x = "Sepal Length", y = "Sepal Width") +
    ggplot2::facet_grid(~ Species) +
    ddplot::theme_dereck()

data.frame(year = time(nhtemp), tmps = as.matrix(nhtemp)) |>
    ggplot2::ggplot(ggplot2::aes(x = year, y = tmps)) +
    ggplot2::geom_line() +
    # geom_hline(yintercept = 0, size = 1, colour = "#333333") +
    ggplot2::labs(title = "New Hampshire yearly temperatures", subtitle = "Collected between 1912 and 1971", x = "Year", y = "Average temperature") +
    ddplot::theme_dereck()

}
\author{
Dereck de Mezquita

This function allows you to add the derecksnotes theme to your ggplot graphics.
}
\keyword{theme_derecksnotes}
