% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{to.data.frame}
\alias{to.data.frame}
\title{Convert a data.table to a data.frame}
\usage{
to.data.frame(x, id.col = NULL, drop.id.col = TRUE)
}
\arguments{
\item{x}{A data.table object to convert.}

\item{id.col}{The name of the column to use as row names. If NULL, no row names are set. Default is NULL.}

\item{drop.id.col}{Logical. If TRUE and id.col is specified, the id column is removed after setting row names. Default is TRUE.}
}
\value{
A data.frame converted from the input data.table.
}
\description{
This function converts a data.table to a data.frame, optionally setting row names from a specified column.
}
\examples{
library(data.table)
dt <- data.table(id = letters[1:3], value = 1:3)
to.data.frame(dt, id.col = "id")
# Returns a data.frame with row names set to "a", "b", "c" and "id" column removed
}
