% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{valueCoordinates}
\alias{valueCoordinates}
\title{Find coordinates of a specific value in an object}
\usage{
valueCoordinates(obj, value = NA)
}
\arguments{
\item{obj}{The object to find the value in. Should be a matrix or data frame.}

\item{value}{The value to find in the object. Default is NA.}
}
\value{
A data frame with the column and row indices of the value in the object.
}
\description{
This function finds the coordinates (row and column indices) of a specific value in an object.
If no value is provided, it finds the coordinates of NA values.
}
\examples{
mat <- matrix(1:9, nrow = 3)
mat[2, 2] <- NA
valueCoordinates(mat)
# Returns a data frame with the coordinates of NA
valueCoordinates(mat, 5)
# Returns a data frame with the coordinates of the value 5
}
