% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparisons.R
\name{Comparison}
\alias{Comparison}
\title{Comparison Class}
\description{
The `Comparison` class represents a comparison between two groups of samples. It includes methods to initialize, print, and validate the data.
}
\details{
An R6 class that represents a comparison between two groups of samples. 
This class contains the comparison name, the group order, and the comparison table.
}
\section{Initialization}{

```
Comparison$new(
  comparison_name = "My comparison",
  group_order = c("Control group", "Test group"),
  comparison_table = data.table::data.table(
       group = c("Control group", "Control group", "Test group", "Test group"),
       sample = c("Sample1", "Sample2", "Sample3", "Sample4")
   )
)
```
}

\section{Usage}{

```
comparison <- Comparison$new(
  comparison_name = "My comparison",
  group_order = c("Control group", "Test group"),
  comparison_table = data.table::data.table(
       group = c("Control group", "Control group", "Test group", "Test group"),
       sample = c("Sample1", "Sample2", "Sample3", "Sample4")
   )
)
comparison$print()  # Print the Comparison object
```
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{comparison_name}}{Character. The name of the comparison.}

\item{\code{group_order}}{Character vector. The order of groups for the comparison, with length 2. The first element will be treated as the "control" group, and the second as the "test" group.}

\item{\code{comparison_table}}{A data.table that contains the group and sample information for the comparison.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Comparison-new}{\code{Comparison$new()}}
\item \href{#method-Comparison-print}{\code{Comparison$print()}}
\item \href{#method-Comparison-clone}{\code{Comparison$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comparison-new"></a>}}
\if{latex}{\out{\hypertarget{method-Comparison-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comparison$new(comparison_name, group_order, comparison_table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comparison_name}}{A character string representing the name of the comparison. Must be of length 1 and not exceed 100 characters.}

\item{\code{group_order}}{A character vector specifying the order of groups for the comparison. Must be of length 2.}

\item{\code{comparison_table}}{A data.table that contains group and sample information for the comparison. This table is crucial for tracking which samples belong to which clinical group. It is used to distinguish samples for the purposes of experimental analysis. It should have two columns, named "group" and "sample". The "group" column identifies the clinical group to which each sample belongs, and the "sample" column lists the names/IDs of the samples. The groups in this table should match the names specified in `group_order`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comparison-print"></a>}}
\if{latex}{\out{\hypertarget{method-Comparison-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comparison$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comparison-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Comparison-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comparison$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
