% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{monte_carlo}
\alias{monte_carlo}
\title{Monte Carlo Simulation}
\usage{
monte_carlo(seed_price, daily_vol, num_sims, num_days)
}
\arguments{
\item{seed_price}{The starting price}

\item{daily_vol}{The daily volatility}

\item{num_sims}{The number of simulations to run}

\item{num_days}{The number of days to simulate}
}
\value{
A list containing the simulation results and end prices
}
\description{
Monte Carlo Simulation
}
