% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_movingaverages.R
\name{stat_movingaverages}
\alias{stat_movingaverages}
\title{Moving averages `ggplot2` layer}
\usage{
stat_movingaverages(
  mapping = NULL,
  data = NULL,
  geom = "line",
  position = "identity",
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE,
  size = 1.75,
  alpha = 0.75,
  FUN = NULL,
  n = list(short = 20L, long = 200L),
  colours = list(short = "yellow", long = "purple"),
  wilder = NA,
  ...
)
}
\arguments{
\item{mapping}{A `ggplot2::aes` object (required - default `NULL`).}

\item{data}{A `data.table` object (required - default `NULL`).}

\item{size}{A `numeric` vector of length one; the size of the line (optional - default `1.75`).}

\item{alpha}{A `numeric` vector of length one; the alpha of the line (optional - default `0.75`).}

\item{FUN}{A `function` to calculate the moving average (can be simple/exponential or any algorithm). The only requirement is that the function a `numeric` vector as a result. The interval value is defined as a secondary argument to this stat see argument `n` (required - default `NULL`).}

\item{n}{A `list` with two elements "short" and "long". These are the intervals for the short and long moving averages (note if these are longer than data points provided you see an error) (required - default `list(short = 20L, long = 200L)`).}

\item{colours}{A `list` with three elements "short" and "long". These are the colours for the short and long moving averages (optional - default `list(short = "red", long = "blue")`).}

\item{wilder}{A `logical` of length one; if set this argument is passed on to the `FUN` which should be able to accept it - for choosing between a classic EMA or a Wilder's EMA (optional - default `NA`).}
}
\value{
A `ggplot2::layer` object.
}
\description{
`stat_movingaverages` is a `ggplot2` layer that allows you to plot moving averages on a `ggplot2` plot either by providing a function to calculate the moving averages or by providing the column names `ggplot2::aes` of the previously calculated moving averages.

You are free to use whatever algorithm you desire, as long as:

1. If using a function it must return a `numeric` vector of the same length as the input vector and accept the arguments `x` (`numeric` vector of value) `n` (`numeric` interval of time used to calculate the moving average); an optional `wilder` argument applicable to `ema` type functions who use the wilder smoothing method.

2. If using previously calculated moving averages, you must provide them as `ggplot2::aes` values; the `aes` must be `short` and `long`.

See examples for more details.
}
\details{
This is a `ggplot2` extension; it is used with the `+` operator for adding a layer to a `ggplot2` object.
}
\section{Aesthetics}{

\code{stat_movingaverages} understands the following aesthetics (required aesthetics are in bold):
\itemize{
  \item \strong{x} -- datetime (x-axis)
  \item y -- required if using `FUN` to calculate moving averages (y-axis)
  \item short -- required if not using `FUN` to calculate moving averages (y-axis)
  \item long -- required if not using `FUN` to calculate moving averages (y-axis)
}
}

\examples{

# get some financial data
# kucoin is private package - you can use any data source
ticker <- "BTC/USDT"

dt <- kucoin::get_market_data(
    symbols = ticker,
    from = "2022-11-28 15:29:43 EST", # lubridate::now() - lubridate::days(7),
    to = "2022-12-05 15:29:31 EST",# lubridate::now(),
    frequency = "1 hour"
)

dt

# we need a function that calculates the indicator for us
# typically I like to write my own functions in C++; in this case we will use TTR's
# the stat expects a named list to be returned - we redefine ttr
bb <- function(close, n = 2, sd = 2) {
    return(as.list(as.data.frame(TTR::BBands(close, n = n, sd = sd))))
}

ema <- function(close, n = 2, wilder = TRUE) {
    return(as.list(as.data.frame(TTR::EMA(close, n = n, wilder = wilder))))
}

dt |>
    ggplot2::ggplot(ggplot2::aes(
        x = datetime,
        open = open,
        close = close,
        high = high,
        low = low,
        group = symbol
    )) +
    ## ------------------------------------
    ddplot::stat_candlestick() +
    ## ------------------------------------
    # ddplot::stat_bollingerbands(ggplot2::aes(y = close), FUN = bb, alpha = list(mavg = 0.5, ribbon = 0.25)) +
    ddplot::stat_movingaverages(ggplot2::aes(y = close), FUN = ema, n = list(short = 10, long = 50), alpha = list(mavg = 0.5)) +
    ## ------------------------------------
    ggplot2::scale_x_continuous(n.breaks = 25, labels = \(x) {
        lubridate::floor_date(lubridate::as_datetime(x), "hours")
    }) +
    ggplot2::scale_y_continuous(n.breaks = 25) +
    ggplot2::labs(
        title = ticker,
        x = "Date",
        y = "Price (USD)"
    ) +
    ddplot::theme_dereck_dark() +
    ggplot2::theme(
        axis.text.x = ggplot2::element_text(angle = 75, vjust = 0.925, hjust = 0.975),
        panel.grid.minor = ggplot2::element_blank()
    )

## ------------------------------------
# you can also provide the column names as aes instead of calculating them by the passing of a function

# calculate the short and long moving averages
dt[, ema_short := ema(close, n = 10, wilder = TRUE)]
dt[, ema_long := ema(close, n = 50, wilder = TRUE)]

dt |>
    ggplot2::ggplot(ggplot2::aes(
        x = datetime,
        open = open,
        close = close,
        high = high,
        low = low,
        group = symbol
    )) +
    ## ------------------------------------
    ddplot::stat_candlestick() +
    ## ------------------------------------
    # provide the colnames to the calculated indicators as aes values
    ddplot::stat_movingaverages(ggplot2::aes(short = ema_short, long = ema_long), alpha = list(mavg = 0.5)) +
    ## ------------------------------------
    ggplot2::scale_x_continuous(n.breaks = 25, labels = \(x) {
        lubridate::floor_date(lubridate::as_datetime(x), "hours")
    }) +
    ggplot2::scale_y_continuous(n.breaks = 25) +
    ggplot2::labs(
        title = ticker,
        x = "Date",
        y = "Price (USD)"
    ) +
    ddplot::theme_dereck_dark() +
    ggplot2::theme(
        axis.text.x = ggplot2::element_text(angle = 75, vjust = 0.925, hjust = 0.975),
        panel.grid.minor = ggplot2::element_blank()
    )

}
\author{
Dereck de Mezquita
}
